+++ src/s_cockpit/assets/__init__.py
@@ -1,0 +1,1 @@
+"""S-Cockpit browser assets."""