+++ tests/__init__.py
@@ -1,1 +1,1 @@
-"""Personal Cockpit tests."""
+"""S-Cockpit tests."""